PRSD-NONE: Moves AlwaysTrueValidator into mockObjects
