+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/forms/journeys/LandlordRegistrationJourneyTests.kt
@@ -24,8 +24,8 @@ import uk.gov.communities.prsdb.webapp.services.EmailNotificationService
 import uk.gov.communities.prsdb.webapp.services.JourneyDataService
 import uk.gov.communities.prsdb.webapp.services.LandlordService
 import uk.gov.communities.prsdb.webapp.testHelpers.builders.JourneyDataBuilder
+import uk.gov.communities.prsdb.webapp.testHelpers.mockObjects.AlwaysTrueValidator
 import uk.gov.communities.prsdb.webapp.testHelpers.mockObjects.MockLandlordData
-import uk.gov.communities.prsdb.webapp.testHelpers.validators.AlwaysTrueValidator
 import java.net.URI
 
 class LandlordRegistrationJourneyTests {
