+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/forms/journeys/PropertyRegistrationJourneyTests.kt
@@ -28,8 +28,8 @@ import uk.gov.communities.prsdb.webapp.services.LandlordService
 import uk.gov.communities.prsdb.webapp.services.LocalAuthorityService
 import uk.gov.communities.prsdb.webapp.services.PropertyRegistrationService
 import uk.gov.communities.prsdb.webapp.testHelpers.builders.JourneyDataBuilder
+import uk.gov.communities.prsdb.webapp.testHelpers.mockObjects.AlwaysTrueValidator
 import uk.gov.communities.prsdb.webapp.testHelpers.mockObjects.MockLandlordData
-import uk.gov.communities.prsdb.webapp.testHelpers.validators.AlwaysTrueValidator
 import java.net.URI
 
 class PropertyRegistrationJourneyTests {
